requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.41.0
httpx==0.28.1
redis==7.3.0
asyncpg==0.31.0
sqlalchemy[asyncio]==2.0.48
alembic==1.18.4
pydantic==2.12.5
pydantic-settings==2.13.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.22
jinja2==3.1.6
aiohttp==3.13.3
python-dotenv==1.2.2
pytest==9.0.2
pytest-asyncio==1.3.0
